Change double to float to reduce memory size

+++ ConvNetCS/Network/ILossLayer.cs
@@ -4,8 +4,8 @@ namespace ConvNetCS
 {
     public interface ILossLayer
     {
-        double Backward(double[] y);
-        double Backward(int y);
+        float Backward(float[] y);
+        float Backward(int y);
         Vol Forward(Vol V, bool is_training);
         List<ParamsAndGrads> getParamsAndGrads();
         Vol Output { get; set; }
