+++ ConvNetCS/QLearn/Window.cs
@@ -8,27 +8,27 @@ namespace ConvNetCS
 {
     public class Window
     {
-        public Queue<double> V { get; set; }
+        public Queue<float> V { get; set; }
         public int Size { get; set; }
         public int minsize { get; set; }
-        public double sum { get; set; }
+        public float sum { get; set; }
         public Window()
         {
-            this.V = new Queue<double>();
+            this.V = new Queue<float>();
             this.Size = 100;
             this.minsize = 20;
             this.sum = 0;
         }
         public Window(int size, int minsize)
         {
-            this.V = new Queue<double>();
+            this.V = new Queue<float>();
             this.Size = size;
             this.minsize = minsize;
             this.sum = 0;
         }
 
 
-        public void Add(double x)
+        public void Add(float x)
         {
             this.V.Enqueue(x);
             sum += x;
@@ -38,15 +38,15 @@ namespace ConvNetCS
                 this.sum -= xold;
             }
         }
-        public double get_average()
+        public float get_average()
         {
             if (this.V.Count < this.minsize) return -1;
-            else return this.sum / (double)this.V.Count;
+            else return this.sum / (float)this.V.Count;
         }
 
         public void reset()
         {
-            this.V = new Queue<double>(); 
+            this.V = new Queue<float>(); 
             this.sum = 0;
         }
     }
